+++ crates/gpui_platform/src/lib.rs
@@ -1,1 +1,3 @@
 pub use current_platform::*;
+#[cfg(any(test, feature = "test-support"))]
+mod tests;
